split nested_schema_adapter.rs into smaller modules—e.g. factory.rs, adapter.rs, and mapping.rs—to improve navigability and reduce merge conflicts.
