+++ datafusion/datasource/src/nested_schema_adapter/tests.rs
@@ -15,187 +15,6 @@
 // specific language governing permissions and limitations
 // under the License.
 
-//! [`NestedStructSchemaAdapter`] and [`NestedStructSchemaAdapterFactory`] to adapt file-level record batches to a table schema.
-//!
-//! Adapter provides a method of translating the RecordBatches that come out of the
-//! physical format into how they should be used by DataFusion.  For instance, a schema
-//! can be stored external to a parquet file that maps parquet logical types to arrow types.
-
-use crate::schema_adapter::{
-    create_field_mapping, SchemaAdapter, SchemaAdapterFactory, SchemaMapper,
-    SchemaMapping,
-};
-use arrow::{
-    array::{Array, ArrayRef, StructArray},
-    compute::cast,
-    datatypes::{DataType::Struct, Field, Schema, SchemaRef},
-};
-use datafusion_common::{arrow::array::new_null_array, Result};
-use std::sync::Arc;
-
-#[cfg(test)]
-use {
-    crate::schema_adapter::DefaultSchemaAdapterFactory, arrow::record_batch::RecordBatch,
-    datafusion_common::ColumnStatistics,
-};
-/// Factory for creating [`NestedStructSchemaAdapter`]
-///
-/// This factory creates schema adapters that properly handle schema evolution
-/// for nested struct fields, allowing new fields to be added to struct columns
-/// over time.
-#[derive(Debug, Clone, Default)]
-pub struct NestedStructSchemaAdapterFactory;
-
-impl SchemaAdapterFactory for NestedStructSchemaAdapterFactory {
-    fn create(
-        &self,
-        projected_table_schema: SchemaRef,
-        table_schema: SchemaRef,
-    ) -> Box<dyn SchemaAdapter> {
-        Box::new(NestedStructSchemaAdapter::new(
-            projected_table_schema,
-            table_schema,
-        ))
-    }
-}
-
-impl NestedStructSchemaAdapterFactory {
-    /// Create a new factory for mapping batches from a file schema to a table
-    /// schema with support for nested struct evolution.
-    ///
-    /// This is a convenience method that handles nested struct fields properly.
-    pub fn from_schema(table_schema: SchemaRef) -> Box<dyn SchemaAdapter> {
-        Self.create(Arc::clone(&table_schema), table_schema)
-    }
-
-    /// Determines if a schema contains nested struct fields that would benefit
-    /// from special handling during schema evolution
-    pub fn has_nested_structs(schema: &Schema) -> bool {
-        schema
-            .fields()
-            .iter()
-            .any(|field| matches!(field.data_type(), Struct(_)))
-    }
-}
-
-/// A SchemaAdapter that handles schema evolution for nested struct types
-#[derive(Debug, Clone)]
-pub struct NestedStructSchemaAdapter {
-    /// The schema for the table, projected to include only the fields being output (projected) by the
-    /// associated ParquetSource
-    projected_table_schema: SchemaRef,
-    /// The entire table schema for the table we're using this to adapt.
-    ///
-    /// This is used to evaluate any filters pushed down into the scan
-    /// which may refer to columns that are not referred to anywhere
-    /// else in the plan.
-    table_schema: SchemaRef,
-}
-
-impl NestedStructSchemaAdapter {
-    /// Create a new NestedStructSchemaAdapter with the target schema
-    pub fn new(projected_table_schema: SchemaRef, table_schema: SchemaRef) -> Self {
-        Self {
-            projected_table_schema,
-            table_schema,
-        }
-    }
-
-    pub fn projected_table_schema(&self) -> &Schema {
-        self.projected_table_schema.as_ref()
-    }
-
-    pub fn table_schema(&self) -> &Schema {
-        self.table_schema.as_ref()
-    }
-}
-
-impl SchemaAdapter for NestedStructSchemaAdapter {
-    fn map_column_index(&self, index: usize, file_schema: &Schema) -> Option<usize> {
-        let field_name = self.table_schema.field(index).name();
-        file_schema.index_of(field_name).ok()
-    }
-
-    fn map_schema(
-        &self,
-        file_schema: &Schema,
-    ) -> Result<(Arc<dyn SchemaMapper>, Vec<usize>)> {
-        let (field_mappings, projection) = create_field_mapping(
-            file_schema,
-            &self.projected_table_schema,
-            |file_field, table_field| {
-                // Special handling for struct fields - always include them even if the
-                // internal structure differs, as we'll adapt them later
-                match (file_field.data_type(), table_field.data_type()) {
-                    (Struct(_), Struct(_)) => Ok(true),
-                    _ => {
-                        // For non-struct fields, use the regular cast check
-                        crate::schema_adapter::can_cast_field(file_field, table_field)
-                    }
-                }
-            },
-        )?;
-
-        Ok((
-            Arc::new(SchemaMapping::new(
-                Arc::clone(&self.projected_table_schema),
-                field_mappings,
-                Arc::new(|array: &ArrayRef, field: &Field| adapt_column(array, field)),
-            )),
-            projection,
-        ))
-    }
-}
-
-// Helper methods for nested struct adaptation
-/// Adapt a column to match the target field type, handling nested structs specially
-fn adapt_column(source_col: &ArrayRef, target_field: &Field) -> Result<ArrayRef> {
-    match target_field.data_type() {
-        Struct(target_fields) => {
-            // For struct arrays, we need to handle them specially
-            if let Some(struct_array) = source_col.as_any().downcast_ref::<StructArray>()
-            {
-                // Create a vector to store field-array pairs with the correct type
-                let mut children: Vec<(Arc<Field>, Arc<dyn Array>)> = Vec::new();
-                let num_rows = source_col.len();
-
-                // For each field in the target schema
-                for target_child_field in target_fields {
-                    // Create Arc<Field> directly (not Arc<Arc<Field>>)
-                    let field_arc = Arc::clone(target_child_field);
-
-                    // Try to find corresponding field in source
-                    match struct_array.column_by_name(target_child_field.name()) {
-                        Some(source_child_col) => {
-                            // Field exists in source, adapt it
-                            let adapted_child =
-                                adapt_column(source_child_col, target_child_field)?;
-                            children.push((field_arc, adapted_child));
-                        }
-                        None => {
-                            // Field doesn't exist in source, add null array
-                            children.push((
-                                field_arc,
-                                new_null_array(target_child_field.data_type(), num_rows),
-                            ));
-                        }
-                    }
-                }
-
-                // Create new struct array with all target fields
-                let struct_array = StructArray::from(children);
-                Ok(Arc::new(struct_array))
-            } else {
-                // Not a struct array, but target expects struct - return nulls
-                Ok(new_null_array(target_field.data_type(), source_col.len()))
-            }
-        }
-        // For non-struct types, just cast
-        _ => Ok(cast(source_col, target_field.data_type())?),
-    }
-}
-
-#[cfg(test)]
 mod tests {
     use super::*;
     use arrow::{
